some parts are working

+++ Includes/SeedCreators/RandStrobeCreators/RandStrobeCreatorGuoPibri.hpp
@@ -7,7 +7,7 @@ class RandStrobeCreatorGuoPibri : public RandStrobeCreator
 {
 	public:
 		RandStrobeCreatorGuoPibri(Hasher* hasher, Comparator* comparator, size_t kmer_len, size_t w_min, size_t w_max,
-				uint8_t n = DEFAULT_N, uint64_t mask = DEFAULT_MASK);
+				uint32_t n = DEFAULT_N, uint64_t mask = DEFAULT_MASK);
 
 		std::vector<Seed*> create_seeds(const std::string& seq, const std::vector<uint64_t>& kmers,
 				const std::vector<uint64_t> hashes);
